moved logic into \integration\appeng\ModAppEng.scala
